feat: add some benchmarks

+++ zero/zero_mq.py
@@ -76,6 +76,10 @@ class ZeroMQPythonDevice(ZeroMQInterface):
         worker_ipc: str,
         worker_port: int,
     ):
+        ctx: zmq.Context = None  # type: ignore
+        gateway: zmq.Socket = None  # type: ignore
+        backend: zmq.Socket = None  # type: ignore
+
         try:
             ctx = zmq.Context.instance()
             gateway = ctx.socket(zmq.ROUTER)
@@ -88,12 +92,8 @@ class ZeroMQPythonDevice(ZeroMQInterface):
             else:
                 backend.bind(f"tcp://127.0.0.1:{worker_port}")
 
+            # details: https://learning-0mq-with-pyzmq.readthedocs.io/en/latest/pyzmq/devices/queue.html
             zmq.device(zmq.QUEUE, gateway, backend)
-
-            gateway.close()
-            backend.close()
-            ctx.destroy()
-            ctx.term()
 
         except KeyboardInterrupt:
             logging.info("Caught KeyboardInterrupt, terminating workers")
@@ -103,10 +103,10 @@ class ZeroMQPythonDevice(ZeroMQInterface):
             logging.error("bringing down zmq device")
 
         finally:
-            gateway.close()
-            backend.close()
-            ctx.destroy()
-            ctx.term()
+            gateway.close() if gateway else None
+            backend.close() if backend else None
+            ctx.destroy() if ctx else None
+            ctx.term() if ctx else None
 
     def worker(
         self,
@@ -115,6 +115,9 @@ class ZeroMQPythonDevice(ZeroMQInterface):
         worker_id: int,
         process_message: Callable,
     ):
+        ctx: zmq.Context = None  # type: ignore
+        socket: zmq.Socket = None  # type: ignore
+
         try:
             ctx = zmq.Context.instance()
             socket: zmq.Socket = ctx.socket(zmq.DEALER)
